+++ changelog.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog][keep-a-changelog]
 ## [Unreleased]
 - Nothing right now
 
+## [0.1.7] (2020-09-02)
+
+### Added
+- Added IDKollen API provider (idkbankid)
+
+### Fixed
+- Added error handling on internal errors when following requests, should create more stable library (all modules).
+- Verified API and updated settings for Funktionstjänster (CGI) Freja eID module (ftfrejaeid)
+
 ## [0.1.5] (2020-09-01)
 
 ### Fixed
@@ -63,6 +72,7 @@ The format is based on [Keep a Changelog][keep-a-changelog]
 
 [keep-a-changelog]: http://keepachangelog.com/en/1.0.0/
 [Unreleased]: https://github.com/DSorlov/eid-provider/compare/master...dev
+[0.1.7]: https://github.com/DSorlov/eid-provider/releases/tag/v0.1.7
 [0.1.5]: https://github.com/DSorlov/eid-provider/releases/tag/v0.1.5
 [0.1.4]: https://github.com/DSorlov/eid-provider/releases/tag/v0.1.4
 [0.1.3]: https://github.com/DSorlov/eid-provider/releases/tag/v0.1.3